Move assertSequenceValues into dedicated test-helpers file
- Extract assertSequenceValues out of live_migration_testing_framework.go
  into a new live_migration_test_helpers.go, gated on the same four test
  build tags. Keeps the framework file focused on the LiveMigrationTest
  state machine.
- Exclude src/testlivemigration from the leaf-error fmt.Errorf lint in
  go.yml. The package's files are test-only but use build tags instead
  of the _test.go suffix, so they don't get picked up by the existing
  *_test.go exclude.
